requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.4.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info==0.18
future==0.17.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
isort==4.3.15
jedi==0.13.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mccabe==0.6.1
mongoengine==0.17.0
netifaces==0.10.4
numpy==1.15.2
PAM==0.4.2
parso==0.3.4
pbr==5.1.3
pluggy==0.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycrypto==2.6.1
pydocstyle==3.0.0
pyflakes==2.1.1
pygobject==3.26.1
pymongo==3.7.2
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.3+ubuntu1
python-bitcoinrpc==1.0
python-debian==0.1.32
python-jsonrpc-server==0.1.2
python-language-server==0.24.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
rope==0.12.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
snowballstemmer==1.2.1
ssh-import-id==5.7
stevedore==1.30.1
systemd-python==234
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.4.3
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.4
wrapt==1.11.1
yapf==0.26.0
zope.interface==4.3.2
